+++ updates/0001/install_python_venv.sh
@@ -1,4 +1,0 @@
-#!/usr/bin/bash
-set -e
-
-sudo apt -y install python3.12-venv